Init of repository
Note that ClassA and ClassB are identical except for the
class name, and both files use tabs (thus not following
the editorconfig). ClassA should be compiled, whereas
the conent of ClassB should be copied to the output.

When applying "Format Document" only ClassA make use of
the editorconfig-settings (ClassB just use the standard
settings is VS).